add a sandbox segment for test purposes
