+++ lib/segment/application.ex
@@ -1,11 +1,13 @@
 defmodule Segment.Application do
   use Application
 
+  @api Application.fetch_env!(:segment, :api)
+
   def start(_type, _args) do
     import Supervisor.Spec, warn: false
 
     children = [
-      {Segment, []}
+      {@api, []}
     ]
 
     opts = [strategy: :one_for_one, name: Segment.Supervisor]
