+++ test/segment/server_test.exs
@@ -1,8 +1,8 @@
-defmodule SegmentTest do
+defmodule Segment.ServerTest do
   use ExUnit.Case, async: false
   use ExVCR.Mock, adapter: ExVCR.Adapter.Hackney
 
-  alias Segment.{Context, Track}
+  alias Segment.{Context, Track, Server}
 
   setup_all do
     ExVCR.Config.cassette_library_dir("test/fixtures/vcr_cassettes")
@@ -11,14 +11,14 @@ defmodule SegmentTest do
     :ok
   end
 
-  describe "Segment.track/4" do
+  describe "Segment.Server.track/4" do
     test "that track sends a request to segment and recieves a 200" do
       ExVCR.Config.filter_request_options("basic_auth")
 
       use_cassette "example_segment" do
-        assert :ok = Segment.send_track("343434", "Test Event")
+        assert :ok = Server.send_track("343434", "Test Event")
         t = %Track{userId: "12345", event: "Test2 Event", properties: %{}, context: Context.new}
-        assert {:noreply, _} = Segment.handle_cast(t, %{})
+        assert {:noreply, _} = Server.handle_cast(t, %{})
       end
     end
   end
